+++ Countly.h
@@ -12,13 +12,14 @@
 #import "CountlyRemoteConfig.h"
 #import "CountlyFeedbackWidget.h"
 #import "CountlyViewTracking.h"
+#import "Resettable.h"
 #if (TARGET_OS_IOS || TARGET_OS_OSX)
 #import <UserNotifications/UserNotifications.h>
 #endif
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface Countly : NSObject
+@interface Countly : NSObject <Resettable>
 
 #pragma mark - Core
 
@@ -804,6 +805,18 @@ NS_ASSUME_NONNULL_BEGIN
  * @discussion App launch time can be recorded only once per app launch. So, second and following calls to this method will be ignored.
  */
 - (void)appLoadingFinished;
+
+/**
+ * Reset the state of SDK that it is not initialized.
+ * @discussion Reset shared instances
+ * @discussion Clear request queue and events queue
+ */
+- (void)halt;
+- (void)halt:(BOOL) clearStorage;
+
+/* Combine all events in event queue into a request and attempt to process stored requests on demand
+ */
+- (void)attemptToSendStoredRequests;
 
 NS_ASSUME_NONNULL_END