+++ CountlyConfig.m
@@ -10,6 +10,10 @@
 @property (nonatomic) NSMutableArray<RCDownloadCallback> *remoteConfigGlobalCallbacks;
 @end
 
+@interface CountlyAPMConfig ()
+- (void)enableAPMInternal:(BOOL)enableAPM;
+@end
+
 @implementation CountlyConfig
 
 //NOTE: Countly features
@@ -29,6 +33,8 @@ CLYFeature const CLYPushNotifications   = @"CLYPushNotifications";
 CLYFeature const CLYCrashReporting      = @"CLYCrashReporting";
 #endif
 
+CountlyAPMConfig *apmConfig = nil;
+CountlySDKLimitsConfig *sdkLimitsConfig = nil;
 
 //NOTE: Device ID options
 NSString* const CLYDefaultDeviceID = @""; //NOTE: It will be overridden to default device ID mechanism, depending on platform.
@@ -51,35 +57,59 @@ CLYDeviceIDType const CLYDeviceIDTypeNSUUID     = @"CLYDeviceIDTypeNSUUID";
 #endif
         self.eventSendThreshold = 100;
         self.storedRequestsLimit = 1000;
-        self.crashLogLimit = 100;
-
-        self.maxKeyLength = 128;
-        self.maxValueLength = 256;
-        self.maxSegmentationValues = 100;
-
+        self.crashLogLimit = kCountlyMaxBreadcrumbCount;
+        
+        self.maxKeyLength = kCountlyMaxKeyLength;
+        self.maxValueLength = kCountlyMaxValueSize;
+        self.maxSegmentationValues = kCountlyMaxSegmentationValues;
+        
         self.location = kCLLocationCoordinate2DInvalid;
-
+        
         self.URLSessionConfiguration = NSURLSessionConfiguration.defaultSessionConfiguration;
-
+        
         self.internalLogLevel = CLYInternalLogLevelDebug;
-
+        
         self.enableOrientationTracking = YES;
         self.enableServerConfiguration = NO;
         self.remoteConfigGlobalCallbacks = NSMutableArray.new;
     }
-
+    
     return self;
 }
 
 -(void)remoteConfigRegisterGlobalCallback:(RCDownloadCallback) callback
 {
     [self.remoteConfigGlobalCallbacks addObject:callback];
+    
 }
 
 
 - (NSMutableArray<RCDownloadCallback> *) getRemoteConfigGlobalCallbacks
 {
     return self.remoteConfigGlobalCallbacks;
+}
+
+- (void)setEnablePerformanceMonitoring:(BOOL)enablePerformanceMonitoring
+{
+    if (apmConfig == nil) {
+        apmConfig = CountlyAPMConfig.new;
+    }
+    [apmConfig enableAPMInternal:enablePerformanceMonitoring];
+    
+}
+
+- (nonnull CountlyAPMConfig *)apm {
+    if (apmConfig == nil) {
+        apmConfig = CountlyAPMConfig.new;
+    }
+    return apmConfig;
+}
+
+- (nonnull CountlySDKLimitsConfig *)sdkInternalLimits {
+    if (sdkLimitsConfig == nil) {
+        sdkLimitsConfig = CountlySDKLimitsConfig.new;
+    }
+    return sdkLimitsConfig;
 }
 
 @end