+++ CountlyPersistency.m
@@ -28,13 +28,14 @@ NSString* const kCountlyServerConfigPersistencyKey = @"kCountlyServerConfigPersi
 
 NSString* const kCountlyCustomCrashLogFileName = @"CountlyCustomCrash.log";
 
+static CountlyPersistency* s_sharedInstance = nil;
+static dispatch_once_t onceToken;
+
 + (instancetype)sharedInstance
 {
     if (!CountlyCommon.sharedInstance.hasStarted)
         return nil;
 
-    static CountlyPersistency* s_sharedInstance = nil;
-    static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{s_sharedInstance = self.new;});
     return s_sharedInstance;
 }
@@ -295,6 +296,21 @@ NSString* const kCountlyCustomCrashLogFileName = @"CountlyCustomCrash.log";
     }
 }
 
+- (void)resetInstance:(BOOL) clearStorage 
+{
+    CLY_LOG_I(@"%s Clear Storage: %d", __FUNCTION__, clearStorage);
+    [CountlyConnectionManager.sharedInstance sendEvents];
+    [self flushEvents];
+    [self clearAllTimedEvents];
+    [self flushQueue];
+    if(clearStorage)
+    {
+        [self saveToFile];
+    }
+    onceToken = 0;
+    s_sharedInstance = nil;
+}
+
 #pragma mark ---
 
 - (void)recordTimedEvent:(CountlyEvent *)event
@@ -358,7 +374,7 @@ NSString* const kCountlyCustomCrashLogFileName = @"CountlyCustomCrash.log";
             [line writeToFile:crashLogFileURL.path atomically:YES encoding:NSUTF8StringEncoding error:&error];
             if (error)
             {
-                CLY_LOG_W(@"Crash Log File can not be created: \n%@", error);
+                CLY_LOG_W(@"%s, Crash Log File can not be created, got error %@", __FUNCTION__, error);
             }
         }
     });
@@ -388,7 +404,7 @@ NSString* const kCountlyCustomCrashLogFileName = @"CountlyCustomCrash.log";
         [NSFileManager.defaultManager removeItemAtURL:crashLogFileURL error:&error];
         if (error)
         {
-            CLY_LOG_W(@"Crash Log File can not be deleted: \n%@", error);
+            CLY_LOG_W(@"%s, Crash Log File can not be deleted, got error %@", __FUNCTION__, error);
         }
     }
 }
@@ -419,7 +435,7 @@ NSString* const kCountlyCustomCrashLogFileName = @"CountlyCustomCrash.log";
             [NSFileManager.defaultManager createDirectoryAtURL:URL withIntermediateDirectories:YES attributes:nil error:&error];
             if (error)
             {
-                CLY_LOG_W(@"Application Support directory can not be created: \n%@", error);
+                CLY_LOG_W(@"%s, Application Support directory can not be created, got error %@", __FUNCTION__, error);
             }
         }
     });